Refactor: move motion redux types into motion file
