+++ src/redux/types/actions/index.ts
@@ -11,7 +11,7 @@ import { MessageActionTypes } from './message';
 import { UserActionTypes } from './user';
 import { WalletActionTypes } from './wallet';
 
-export { RootMotionOperationNames } from './colonyActions';
+export { RootMotionOperationNames } from './motion';
 
 /*
  * Type that represents an action (bare minimum).
