+++ src/redux/types/actions/motion.ts
@@ -1,7 +1,9 @@
 import { BigNumber } from 'ethers/utils';
+import { ColonyRole } from '@colony/colony-js';
 
 import { ActionTypes } from '~redux/index';
 import { Address } from '~types/index';
+import { Color } from '~core/ColorTag';
 
 import {
   ErrorActionType,
@@ -9,6 +11,11 @@ import {
   ActionTypeWithMeta,
   MetaWithHistory,
 } from './index';
+
+export enum RootMotionOperationNames {
+  MINT_TOKENS = 'mintTokens',
+  UPGRADE = 'upgrade',
+}
 
 export type MotionActionTypes =
   | UniqueActionType<
@@ -88,5 +95,119 @@ export type MotionActionTypes =
   | ErrorActionType<ActionTypes.COLONY_MOTION_CLAIM_ERROR, object>
   | ActionTypeWithMeta<
       ActionTypes.COLONY_MOTION_CLAIM_SUCCESS,
+      MetaWithHistory<object>
+    >
+  | UniqueActionType<
+      ActionTypes.COLONY_MOTION_DOMAIN_CREATE_EDIT,
+      {
+        colonyAddress: Address;
+        colonyName?: string;
+        domainName: string;
+        domainColor?: Color;
+        domainPurpose?: string;
+        annotationMessage?: string;
+        parentId?: number;
+        domainId?: number;
+        isCreateDomain: boolean;
+      },
+      MetaWithHistory<object>
+    >
+  | ErrorActionType<ActionTypes.COLONY_MOTION_DOMAIN_CREATE_EDIT_ERROR, object>
+  | ActionTypeWithMeta<
+      ActionTypes.COLONY_MOTION_DOMAIN_CREATE_EDIT_SUCCESS,
+      MetaWithHistory<object>
+    >
+  | UniqueActionType<
+      ActionTypes.COLONY_MOTION_EXPENDITURE_PAYMENT,
+      {
+        colonyAddress: Address;
+        colonyName?: string;
+        recipientAddress: Address;
+        domainId: number;
+        singlePayment: {
+          amount: BigNumber;
+          tokenAddress: Address;
+          decimals: number;
+        };
+        annotationMessage?: string;
+      },
+      MetaWithHistory<object>
+    >
+  | ErrorActionType<ActionTypes.COLONY_MOTION_EXPENDITURE_PAYMENT_ERROR, object>
+  | ActionTypeWithMeta<
+      ActionTypes.COLONY_MOTION_EXPENDITURE_PAYMENT_SUCCESS,
+      MetaWithHistory<object>
+    >
+  | UniqueActionType<
+      ActionTypes.COLONY_MOTION_EDIT_COLONY,
+      {
+        colonyAddress: Address;
+        colonyName: string;
+        colonyDisplayName: string;
+        colonyAvatarImage?: string;
+        colonyAvatarHash?: string;
+        hasAvatarChanged?: boolean;
+        colonyTokens?: Address[];
+        annotationMessage?: string;
+        /*
+         * @TODO I think this will also store the subscribed-to tokens list
+         */
+      },
+      MetaWithHistory<object>
+    >
+  | ErrorActionType<ActionTypes.COLONY_MOTION_EDIT_COLONY_ERROR, object>
+  | ActionTypeWithMeta<
+      ActionTypes.COLONY_MOTION_EDIT_COLONY_SUCCESS,
+      MetaWithHistory<object>
+    >
+  | UniqueActionType<
+      ActionTypes.COLONY_MOTION_MOVE_FUNDS,
+      {
+        colonyAddress: Address;
+        colonyName?: string;
+        tokenAddress: Address;
+        fromDomainId: number;
+        toDomainId: number;
+        amount: BigNumber;
+        annotationMessage?: string;
+      },
+      MetaWithHistory<object>
+    >
+  | ErrorActionType<ActionTypes.COLONY_MOTION_MOVE_FUNDS_ERROR, object>
+  | ActionTypeWithMeta<
+      ActionTypes.COLONY_MOTION_MOVE_FUNDS_SUCCESS,
+      MetaWithHistory<object>
+    >
+  | UniqueActionType<
+      ActionTypes.COLONY_ROOT_MOTION,
+      {
+        operationName: RootMotionOperationNames;
+        colonyAddress: Address;
+        colonyName?: string;
+        motionParams: [BigNumber] | [string];
+        annotationMessage?: string;
+      },
+      MetaWithHistory<object>
+    >
+  | ErrorActionType<ActionTypes.COLONY_ROOT_MOTION_ERROR, object>
+  | ActionTypeWithMeta<
+      ActionTypes.COLONY_ROOT_MOTION_SUCCESS,
+      MetaWithHistory<object>
+    >
+  | UniqueActionType<
+      ActionTypes.COLONY_MOTION_USER_ROLES_SET,
+      {
+        colonyAddress: Address;
+        colonyName: string;
+        domainId: number;
+        userAddress: Address;
+        roles: Record<ColonyRole, boolean>;
+        annotationMessage?: string;
+      },
+      MetaWithHistory<object>
+    >
+  | ErrorActionType<ActionTypes.COLONY_MOTION_USER_ROLES_SET_ERROR, object>
+  | ActionTypeWithMeta<
+      ActionTypes.COLONY_MOTION_USER_ROLES_SET_SUCCESS,
       MetaWithHistory<object>
     >;
